rfc2: allow [L]GPLv2+ licensing; wordsmithing
Given that LLNL has not yet approved [L]GPLv3 code release, and we are
about to release Flux components, we will have to do so under [L]GPLv2+.
Change the rules to encourage [L]GPLv3+ but allow [L]GPLv2+.

Require the "or later" clause for [L]GPL version so that we can upgrade
the base version of our software later if we want to / are allowed.

Clarify description of Flux user interfaces.