rfc5: add keepalive monitoring, expand
Add an introduction and define some common terms:
base component, extension module, and two styles of
extension module: plugin style (like sched plugins),
and messaging actor style (like comms modules).

Mention the shutdown message used to tell a module it's time
to exit its mod_main().  This is already implemented in
the broker and has been used in a few comms modules.

Add keepalive status reporting and sections decribing
loading, monitoring, and unloading of modules.
flux module list should be able to report these keepalive
states.  The ready/busy states were intended to map to
prepare and check watchers in the Flux reactor.

When loading a module with insmod, wait for state transition
out of initializing before returning a response.  This allows
us to return module initialization errors to flux module load
and the user.  It also allows tests or real situations that load
modules across a session and require a module to complete some
initialization before proceeding to have a way to wait on that.
Possibly this could could be extended to an instance wide global
state that could gate the starting of the initial program,
or be used in other situations.

When unloading a module, replace the current "empty message"
handshake indicating comms module thread exit with a keepalive
transition to state=exited (with possible errnum returned).

Update IETF json-content-rules reference to draft 5,
expires April 21, 2016.