rfc13: describe version 1.0 vs 1.1
Describe what changed in version 1.1.

Require all function signatures to be present in a PMI library,
and identify which functions are OPTIONAL and can be PMI_FAIL
stubs.  1.1 is NOT a shared library major version update.

Add MPI quirks section.  Expect this to be filled in with
more details as we learn more.

More fully describe KVS semantics, and allow the possiblity
of PMI_Barrier() performing the KVS data exchange.

General cleanup.