rfc13: clarify some text passages
Clarify the PMI version description, especially about
the ability for client or server to negotiate down
to version 1.

Since practicaly speaking, use of the protocol as opposed
to the shared library seems to strongly reduce coupling between
process managers and programs, change the language to more
strongly recommend it.  Also, recommend that PMI libraries
be implemented using the wire protocol; that way if an MPI
(for example one not derived from MPICH) happens to get
strongly bound to a PMI library, it can still avoid being bound
to a specific process manager.

Break the long lines in the ABNF.
Drop extra (and illegal) equal signs in spawn ABNF definition.

Make the required library major version 0 not 1.
This is what SLURM explicitly provides.