Add pdf make target for all docs
Add pdf make target that use a2x to convert asciidoc files
into pdf files.

a2x is using docbook, and docbook does not not like empty
sections or sections named "References".  We ad "TBD" to empty
sections to make them non-empty, and explicitly declare the
sections type as "[sect2]" for "== References" sections.

Note that an alternate way of dealing with the "References"
section name problem is to convert to using actual footnotes
and/or bibliography support in asciidoc.