rfc11: add hdir object type
Problem:  KVS performs poorly when directories grow to contain
thousands of entries.  Each update requires the entire directory,
a json dictionary, to be encoded and rewritten as a new object,
and each lookup requires the entire directory to be read and
decoded.

The proposed solution consists of replacing a "dir" treobj
with an "hdir" object once the "dir" exceeds some configurable
number of entries.

The "hdir" object uses a configurable hash function to map key names
to integers which are divided modulo a fixed number of buckets to
obtain a bucket index.  The bucket index references a "dir" or
"hdir" object that will hold the mapping for the requested key.

This strategy has the following properties:
1) The entire namespace for a directory need not be loaded in order
to look up a key.  The hash function allows this operation to
select smaller objects yet still definitively determine whether a
given key is present.
2) The entire namespace for a directory need not be written out as
a new content store object when a new key is added to the directory.
Again, only smaller objects representing a subset of the namespace
are involved in the update.