rfc3: drop "cmb." prefix from broker services
Problem: the required "cmb." prefix on broker-resident
services makes it impossible to migrate the service
to a comms module without changing the protocol.

Drop this requirement.  The broker routing already uses
a service table rather than this tag to determine where
to route a given request message.  Indeed a large number
of broker-resident services have already dropped the cmb prefix.