rfc3: add userid, role mask 32-bit integers to PROTO
Two new 32-bit integer fields are added to the PROTO block of
the CMB/1 message.  Since the PROTO block is the same for all
message types, this adds 64 bits to all messages.

The 32-bit userid is intended to map directly to the host system's
definition for POSIX uid_t, with the value for "userid-unknown"
corresponding to (uid_t)-1 as described for POSIX chown(2).
Since POSIX doesn't specify the size of uid_t, this may require some
remapping on operating systems with 16-bit uid_t or different
signed integer representations.

The 32-bit role mask is intended to support a design similar,
but orthogonal to, Linux kernel capabilities, with a value of zero
indicating "no assigned roles".

Although this is a non-backwards compatible protocol change,
the protocol remains version 1 due to the pre-release status
of flux-core.