rfc4:  Revise the notion of "shared"
Eliminate the old "shared" resource allocation element.  In its place,
commandeer the resource's size element to represent the level of
sharing.

Require each unit of size to be allocated to no more than one job; and
multiple jobs can "share" a resource by being allocated disjoint
portions of the size items.

Introduce a new "shared" resource allocation element that contains an
aggregation of jobs allocated to all of the resource's descendents.
The "shared" element contains just the list of jobs, devoid of any
"size" value.

For example, if one node of a rack is allocated exclusively to job 1
and another node of the same rack is allocated exclusively to job 2,
the rack resource's "shared" value will be a list of jobs 1 and 2.

This allows a scheduler to quickly determine when a parental resource
and all its children can be allocated exclusively to a job.