rfc14: Overhaul examples to better match the canonical jobspec
Overhaul the examples to better match the canonical jobspec, so we
have a foundation to debate and improve the jobspec.

Honor the rule, "The dictionary MUST contain the keys `resources`, `tasks`,
`walltime`, `attrs`, and `version`" by adding missing keys.

Eliminate the assumed "tasks" section in the examples.  Instead directly
include complete examples so that the reader is not asked to make mental
substitutions.

Add "count" to any resource that previously had only a "type".  By canon,
resource SHALL contain both a type and a count.

Fixes #83
Fixes #86