rfc14: Fix JCR to use '$' and '=' for Rules
The JSON Content Rules (JCR) 0.6 says:

"Rule names are signified with the dollar character ('$')"

so we update all of the rule names to include the dollar character.

Also, it says:

"Rule name assignments to primitive type specifications separate the rule
name from the type specification with the character sequence '=:', whereas
rule name assignments for all other type specifications only require the
separation using the '=' character."

Our rule name assignments only have a space, so we update all of them to
use the '=' character.