rfc14: Fix location of Kleene symbols in JCR section
Fix the location of the Kleene symbols (e.g. +) in the JSON
Content Rules (JCR) section of rfc14.

The JCR 0.6 says:

"Evaluation of subordinate components in array, object, and group
specifications may be succeeded by a repetition expression denoting
how many times the subordinate component should be evaluated."

Note that is says _succeeded_, but our usage has preceeded with the
symbol.  We move the symbol to match the spec.