rfc14: Remove walltime and expand attributes definition
First, we rename "attrs" to "attributes".  Very little abbreviation
of terms exists elsewhere in the jobspec, and we have explicitly designed
for verbosity over brevity.

Next, we remove the "walltime" from the group of valid top-level keys.
Job duration will instead be expressed as a "system" attribute.

The "attributes" section is expanded to add sub-dictionaries named
"system" and "user".