rfc14: Remove uuid, attributes, and tags from resource vertex
We remove uuid from the jobspec because its intent was for it to
be used more internally to flux.  Perhaps it would be used in internal
data structures, or perhaps it would also be useful in a resource
specification.  But there is no known use for it in the jobspec at
this time.

It seems a bit redundant to have both attributes and tags per resource
vertex.  For instance, a nil value for a key in the attributes would
seem to serve the same purpose as a tag.  But at this time, there is
no clear use case for either of these sections.  Perhaps these
too had more internal uses planned?  We remove them both for now.
The simpler we can make the jobspec definition, the easier it will be
to implement that interacts with jobspecs.

We can always add them in the future if/when use cases are identified.