rfc14: Remove 'group' resource vertex type
Remove the 'group' resource vertex type.  This probably made sense
in the past, but with the current structure of the jobspec it is
not clear that an _unnamed_ grouping serves any purpose.  A use
for a _named_ grouping might be useful, but we will put off adding
that structure until such point as we find a difinitive use for it.