rfc14: Remove "edge" from resource vertex.
We have decided that arbitrary graph resource requests are too
complicated to handle at this time.  We remove the part of the
spec, "edge", that would have allowed the arbitrary graphs that we
cannot support.

If we decide to have edges in the future, we believe that a separate
section for edges is probably the way to go, rather than putting them
under resource vertices.