rfc14: Remove implicit slots
We have found that if we remove implicit slots from the jobspec
it makes implementation more straight forward in places.  For instance,
the scheduler can ignore the "tasks" section of the jobspec and just
focus on the "resources" section.

There is also an arugment to be made that having only one explicit
way to specify slots better honors our goal to "Emphasize expressivity
over simplicity".