rfc15: remove reference to R, modify input to the IMP
Replace the concept of a signed request 'R' from the instance owner
with simply cleartext "input to the IMP." Remove other security-related
contents associated with 'R' from the IMP's input, including Timestamp,
TTL, and UUID, since these no longer make sense.

Add a UID/username field to 'J' that denotes the creator of 'J'.

Add the job shell path to 'J' so the user may select a different job
shell than the default (for testing or specific use cases)

Replace the global assigned resource set that was in 'R' with a local
assigned resource set, computed by the instance.

Add "options" fields to both 'J' and the resource owners input to the
IMP. This allows both the instance and the guest user to control
exposed optional behavior in the IMP or its plugins at runtime.

Reword "Overall Design" section based on changes above.

Reword the "Overall Design" section based on these changes.