rfc14: change static references to asciidoc footnotes
Turn the references in RFC14 into asciidoc footnotes so that the
footnote numbers are clickable.

Also update the URLs for the two changes references to point to more
static locations for the referenced location rather than "current" which
makes the linked-to page change over time and not match the version listed
in our document.