rfc2: clarify goal of license compat with others
Problem: the last item in the "goals" section was poorly
worded, stating that Flux should be usable by software with
incompatible licenses.

Reword that item to be more clear - the point is that Flux
must be licensed in a way that allows it to be widely used
by other software.