spec14: clarify UC1.3, UC2.5 & UC2.6
The description of Use Cases 1.3, 2.5, and 2.5 are worded with "at
least", but the actual example jobspecs specify an exact "count" of
memory, sockets or cores.

Update the example to use a `min` count instead with no `max` (implying
"at least" and making the use case consistent.

Adjust the wording for the task `count` key to allow specification of
`min` without `max`, to make the new UC1.3 adhere to the spec.