rfc21: define events
Add detail regarding the events that the job manager
logs to the primary job eventlog.

Back out the requirement that the event context use
a key=value [key=value...] format, and use JSON instead
so that we don't set ourselves on a path of reinventing
yet another half-baked serialization format.