rfc21: add "and events" to title
Now that this RFC contains detailed information
about the events that trigger job state transitions,
the title is more appropriately "Job States and Events".

Update requirements to show that all state transitions
are triggered by events, and all events are logged to
the job eventlog.  This was not explicit before.

Update the README.adoc index to include the new title
and brief description.