rfc14: fix jsonschema footnote
Problem: The initial footnote referencing JCR (now json-schema)
does not render properly.

I haven't researched this in any great detail, but it seems that
the "footnoteref" reusable footnote style used here is considered
legacy, and a new scheme for reusable footnotes is used by asciidoctor
where footnote:name[text] can be reused with footnote:name[].
However, that didn't seem to render properly on github.

Rather than go down a rat hole on this, I felt the first footnote
was not really needed, and simply eliminated it so there is only
one json-schema footnote reference, thus it doesn't need to be
"resusable".