rfc21: add wait status to finish event context
Problem: the "finish" event doesn't currently
tell the user anything about whether the job
ran OK or not.

Add an integer status code to the finish context,
which will be the largest of the wait(2) status
codes for the job shells.

Since this RFC is describing what happens at the
job manager level, we don't describe job shell behavior
when an application process fails, though one could
argue that the shell(s) should exit with a nonzero exit
status in that case.  This subject MAY be covered in a
future job shell RFC.