rfc14: make duration attribute a number of seconds
Problem: RFC 14 specifies that the system attribute "duration" is
a "string representing a time span", but does not specify how that
string is to be interpreted.

Since this RFC describes a canonical jobspec format, specifically
state that the duration is a number of seconds, leaving no question
as to how it should be interpreted.

Update all use case examples that used arbitrary strings for duration.