rfc24: drop footer event
Problem: the footer event requires some centralized entity
like the leader shell to decode all data events and count
characters, and either be the sole writer, or execute a
barrier across all writers.

Drop the footer event.  It doesn't really add any new
information and it is a pain to generate.