rfc20: add optional "attributes" top-level key
Problem: there is currently no way for a scheduler
to emit opaque information specific to the scheduler itself.
For example, as part of resource restruction work within Fluxion,
a need arose to emit the queue name into RV1 so that
when a multi-queue-capable Fluxion module is reloaded,
each currently running job can be enqueued into
the corresponding queue.

Introduce the "attributes" at the top level which
can have the "system" key to contain system
specific information pertaining to the
creation of this R document.
W/ this change, Fluxion can now stick the queue name
to "attributes.system.scheduler" and fetch this
back to reconstruct its reconstruction correctly.