rfc19: detail FLUID string representations
Expand the "Representation" section to detail the FLUID encodings
which should be supported by a Flux instance, including the existing
representations already supported: decimal, hex, dothex and "words"
(mnemonic encoding).

Additionally, introduce a preferred alternate encoding for FLUIDs
(F58) based on Base58 (using the Bitcoin dictionary) which has
been optimized for human readability by dropping similar letters
and numbers.  (See Bitcoin Wiki reference in the text).

Add a set of "rules" that should be followed to unambiguously decode a
FLUID encoded in one of the standard representations to a 64b integer.