rfc20: replace R_lite "node" key with top-level nodelist
Problem: Encoding hostname in the node key of each R_lite entry of
R version 1 effectively disables any ability to compress the R_lite
array through combining ranks with similar info (unless all hostnames
happend to be the same)

e.g., when adding hostnames to the following simple R_lite:

 {"R_lite":[{"rank":"0-3", "children":{"core":"0-3"}}]}

we get

 {"R_lite":[{"rank":"0", "node":"host0", "children":{"core":"0-3"}},
            {"rank":"1", "node":"host1", "children":{"core":"0-3"}},
            {"rank":"0", "node":"host2", "children":{"core":"0-3"}}]}

Instead, allow a `nodelist` key at the `execution` level of Rv1 as
a way to list the hostname of each rank encoded in `R_lite`.

As an example the above R becomes:

 { "nodelist":["host0","host1","host2"],
   "R_lite":[{"rank":"0-3", "children":{"core":"0-3"}}]
 }

The `nodelist` array may contain entries with embedded idsets, so the
above example could also be represented more compactly as:

 { "nodelist":["host[0-2]"],
   "R_lite":[{"rank":"0-3", "children":{"core":"0-3"}}]
 }

In order to simplify the design of Flux components that deal with Rv1,
make the `nodelist` key mandatory, so that there is always a mapping of
R_lite ranks to hostname for components that wish to display a name in
their output.