rfc21: add PRIORITY state
Add a new PRIORITY state, between DEPEND and SCHED.
A job will remain in PRIORITY state while its priority is
being calculated by a job manager priority plugin.

In many cases this may be instantaneous, but we leave open the
possibility that a plugin may need to contact another service
to obtain user data before assigning a priority.  If this service
is slow or unresponsive, the amount of time spent in PRIORITY state
will provide some visibility into what is going on.

A 'priority' event is posted when the priority is assigned, and
later when the priority is updated.  This event is not posted to
the job eventlog to avoid the extra noise, but would appear in the
job manager journal so that external services like job-info can
reorder their queue.