rfc27: add sched.prioritize RPC
Add an RPC that allows the job manager to send the scheduler
a list of (id, priority) tuples to reprioritize jobs.

When a priority plugin, such as fair share, updates many jobs at
once, all updates can be captured in one request message, so that the
total reordering is received at once by the scheduler.  sched.prioritize
requests may also be sent when single jobs are re-prioritized.

This is an open loop request.  The scheduler does not send a response,
and may ignore unknown job IDs.