rfc21: refine PRIORITY state transitions
Problem: 'priority' event is not persistent and no event
transitions state from SCHED back to PRIORITY on instance
restart.  This causes problems for other components that
recreate job state by replaying the eventlog.

Make the 'priority' event persistent (drop the note that says
otherwise), and change the description to indicate that this
event sets the inital priority, making no reference to updates.

Add a 'flux-restart' event that will transition state from SCHED
back to PRIORITY.  This event is to be posted to each active job
on job manager restart.