rfc27: refactor hello RPC to be streaming
Problem: The original hello protocol sent an array of jobs in a single
RPC response to the scheduler.  This made it difficult to construct
an API for plugin writers that wish to operate on a single job id at
a time.

Solution: Update the the hello RPC to be a streaming RPC, where each
job is streamed to the scheduler.  Users shall wait for ENODATA to
indicate the stream of jobs is complete.