rfc5: remove FLUX_MODSTATE_SLEEPING
Problem: As modules are no longer required to use prepare and check
watchers to transition between SLEEPING/RUNNING states, there is no
longer a practical way for a module to be placed in a SLEEPING state,
and modules are either initializing, running, finalizing, or exited.

Update RFC 5 to remove the FLUX_MODSTATE_SLEEPING state, since it is
no longer applicable.