rfc21: add dependency-add/remove events
Problem: RFC21 requires that job dependencies be managed by a separate
dependency service, but this does not allow simple dependencies to
be handled locally in the job manager.

Add new "dependency-add" and "dependency-remove" events to RFC21 which
indicate when a dependency is added and removed from a job. This allows
more observability and flexibility in the depedency management of a job
and enables the job manager to internally handle simple dependencies.

A separate dependency service can still be used by having the request
to the service issue a `dependency-add` event, and the response post
the `dependency-remove` event.