rfc21: add hold/release-request events
Problem: There is no way in the job manager for a plugin or other
entity to delay the start of a job after the alloc event, nor the
free of job resources after a finish event. However, this would be
useful in numerous cases including global, per-job initialization
or teardown tasks, instance wide job prolog and epilog etc.

Add a new event, `hold-request`, which allow a plugin to block the job
manager from issuing these requests until the plugin releases the
hold by posting a corresponding `release-request` event.