rfc21: add job prolog and epilog events
Problem: There is no way in the job manager for a plugin or other
entity to delay the start of a job after the alloc event, nor the
free of job resources after a finish event. However, this would be
useful in numerous cases including global, per-job initialization or
teardown tasks, instance wide job prolog and epilog etc.

Add a set of new events to be used by these job prolog and epilog
actions: `prolog-start`, `prolog-finish`, `epilog-start`, and
`epilog-finish`.

A `prolog-start` event prevents the job manager from issuing a start
request to the execution system until the corresponding `prolog-finish`
event is posted, and an `epilog-start` event prevents the job manager
from issuing a free request to the scheduler until the corresponding
`epilog-finish` event.

This allows arbitrary, asynchronous work to be done in the job manager
while a job has allocated resources, but before the job either starts
executing, or fully releases those resources to the scheduler.