rfc2: relax LGPL-3.0 requirement
Problem: apparently LGPL becomes viral in the golang environment
because static linking is the best practice there.

If we were to implement a golang binding for Flux, licensing it
as LGPL would defeat one of our stated goals:

* Facilitate the use of Flux and its programming interfaces by external
  projects such as applications, application runtimes, and tools, that are
  distributed under a wide variety of open source and commercial licenses.

Such a project would likely require a permissive license to be useful.
Therefore, relax the LGPL-3.0 REQUIRED language to RECOMMENDED.