rfc20: add optional execution.properties key to Rv1
Problem: Rv1 does not support a method for assigning properties to
resources, but simple property support is a near term requirement for
Flux.

Add an optional properties key to the Rv1 execution dictionary which
allows for the simple assignment of named properties to sets of ranks.

Slightly restrict the set of valid characters in property names.

Reserve the '@' character and any suffix for scheduler-specific use.

Co-authored-by: Jim Garlick <garlick.jim@gmail.com>