rfc25: add optional exclusive key to node resource vertex
Problem: Jobspec V1 does not define an exclusive key for resources
in the resource section. This makes it impossible to offer a consistent
interface for users for the `flux mini` commands when only the `-N`
option is used.

Add `exclusive` key to Jobspec V1 only for `node` type resource vertices.
The intent is to ensure that a user can request exclusive access to nodes
whether an instance allocates nodes exclusively by default or not.