rfc14: add constraints attribute to jobspec
Problem: RFC 31 defines a format for specification of generic job
constraints, but RFC 14 does not define the location for a constraints
in jobspec.

Define attributes.system.constraints as the location in jobspec for
RFC 31 defined constraints.

Update the RFC 14 schema, but only describe `constraints` as an "object",
the actual constraint object is meant to be validated by a scheduler,
since different schedulers may support differing constraint "operators".