rfc20: expand description of scheduler specific properties
Problem: The purpose of the scheduler-specific `@` character in
property strings is vague in RFC 20, and could cause confusion.

Expand the description of the purpose and possible use case for `@`
in execution target properties. Specify that the literal property
string shall still apply to the execution target, but the scheduler
implementation may use all or part of the prefix and suffix before the
`@` for scheduler specific optimizations.