ci: do not run pr-validator action on push
Problem: The pr-validator action is only meant to validate the commits
in a pull request, and can often result in false positive failures when
run in a 'push' (e.g. a push to a user's own fork when that fork's main
branch is not up-to-date).

Only run the pr-validator on 'pull_request'.