rfc28: add property-add, property-remove keys
Problem: There is no way to dynamically update properties in an
instance besides rewriting resource.R and reloading the resource
module and scheduler, but dynamic property updates could be useful
to allow flexible use of resources.

Add property-add and property-remove keys to the `Additional Acquire
Responses` section as a proposed method for allowing the resource
module to update the scheduler with property changes.