rfc23: add "ms" suffix to Flux Standard Duration
Problem: The Flux Standard Duration specification only allows a
unit suffix of "s" (seconds) or greater, but there are times when
it would be convenient to represent a duration in milliseconds (ms),
especially when converting a small duration into FSD.

Add an "ms" suffix to the FSD spec. Reword the specification to allow
for string suffixes instead of only a single character suffix.