rfc21: allow SCHED->PRIORITY on urgency event
Problem: the priority must be recalculated if the job urgency
changes, but since the job is already past PRIORITY state, this
requires special case handling.

In addition, the priority plugin may take a while to calculate
the new priority and this will not be visible to users.  It would
be better to transition back to the PRIORITY state.

Update state diagram to allow for SCHED->PRIORITY transition
on the urgency event.