rfc21: transition NEW->DEPEND on "validate" event
Problem: the current state machine transitions jobs from NEW to
DEPEND on the submit (first) event, which leaves no opportunity for
events to be posted before the job is validated, as may be needed
for assigning default values to empty jobspec attributes such as
queue or duration.

Change the state machine to transition from NEW to DEPEND on a new
"validate" event.  Or if the job is invalid, an "invalidate" event
signifies that the job should immediately be removed.