rfc21: add version to submit event
Problem: the job eventlog format recently changed with the addition
of the validate event, and this made it difficult to detect and
handle eventlogs from earlier versions when restarting after a
Flux upgrade.

Add a version number to the submit event that can be used to
distinguish this and future changes to the eventlog format.

Add "Version 1" to the document title.