rfc3: add message part framing
Problem: rfc3 does not describe the message part framing.

This used to be handed off to ZeroMQ, but now we do the part framing
for all transports, including ZeroMQ.

Describe how message parts are delimited within a message, then
rework the description of the message parts to reflect that they
use this framing, not ZeroMQ's.

Fixes #344