rfc3: fix message framing misinformation
Problem: zeromq handles framing of message parts when it is
performing message transport, so the discussion of message part
framing must be combined with the section on message framing
when a UNIX domain stream socket is used for message transport.

Reorganize the Message Structure and Message Framing and Security
sections, and split the figure into one with all the framing fields,
and one with none.

Fixes #349