rfc3: reorganize document
Problem: RFC 3 doesn't follow a coherent outline.  It was created during
an early period and has been updated without concern for flow.

Reorganize the RFC to hopefully flow in a more coherent manner.

Cut down the ABNF so it only describes the PROTO block (referred to now
as a "message header").

Rework some sections that I found confusing and add various notes and
explanations.