rfc34: document format for unknown task maps
Problem: RFC 34 does not document a task map encoding when the
mapping is unknown, but this is a valid case that must be supported
when the mapping can't be distributed (e.g. the value is too large)
or is otherwise not known.

Document that an empty task map, i.e. a zero-length array, shall
indicate an unknown mapping.

Add an entry to the set of test vectors for the unknown mapping `[]`.