rfc35: add Constraint Query Language RFC
Problem: It would be useful in at least a couple Flux utilities and
programs to support a constraint query syntax which can be used to
generate RFC 31 style JSON constraint objects. However, this mini
language is not specified anywhere.

Attempt to describe a simple, text-based grammar which can be used
to create generic RFC 31 compatible JSON objects for use in any
situation in which these objets may be useful.