rfc36: Add an RFC describing submission directives
Problem: Support for combining job submission options with a submitted
script or file is planned, but there is no specification of how this
will be accomplished.

Add a new RFC describing a specification for the embedding of submission
directives in a batch script or other file.

Update spelling dictionary as required.