rfc24: add repeat key to data event
Problem: when a job produces many identical lines of output,
such as when repeatedly encountering the same error, we have
no way of representing that efficiently in the output eventlog.

Add a repeat key to data events.

Identical data items can be de-duplicated when added to the
eventlog, and re-duplicated when converted back into a stream.