rfc21: allow jobspec-update to any key
Problem: the jobspec-update event is constrained to only
allow updates to the `attributes` section, but updates to the
resources section are needed to implement emerging requirements
in flux-coral2.

Allow any key to be updated.

Add some clarity on JSON key "paths" and what is supposed to happen
if the key already exists.

Fixes #395