rfc28: add expiration to list of possible keys in acquire response
Problem: RFC 21 defines a resource-update event which may modify
the expiration time of the R for a job, but there is no way for
an expiration update of a Flux instance running as a job to be
communicated to the subinstance scheduler.

Add an expiration key to the RFC 28 acquire response which may be
used by the instance reosurce module to notify the scheduler of an
expiration update.