rfc20: rework implementation section
Problem: RFC 20 contains prose that does not elucidate
and gets in the way of using the document as a quick reference.

Drop prose that isn't useful, and avoid needless repetiton.

Add a (type, REQUIRED|OPTIONAL) prelude to each key description
for brevity and consistenency.

Standardize on "execution target" as opposed to "node rank" or
"broker rank" and add a note that explains this choice.

Spell out some things that were implied or vague before:
- R_lite should have at least one entry
- Idsets should have at least one id
- nodelist and R_lite arrays may not have the same number of elements

Add a note about the inability to re-rank the scheduling key.

There were no changes to the description of the attributes key
because there is another PR pending to remove it.