rfc13: define allowed characters in KVS values
Problem: the spec allows '=' and SPACE characters to appear
in KVS values, but some implementations do not allow this.

Recent changes in MPICH hydra caused those characters to no longer
work in a KVS value.  In addition, we observed in the past that the
Cray PMI implementation cannot handle those characters but treated
it as a weakness of their implementation.

Define the allowed characters explicitly.

See also pmodels/mpich#7050.