rfc14: add bank and project as optional jobspec attributes
Problem: The canonical jobspec attributes for `bank` and `project`
are not specified in RFC 14. Even though these attributes only apply
for multi-user instances with an accounting plugin configured, they
are common enough concepts that it would be useful to standardize
their name, location in jobspec, and purpose.

Define `bank` and `project` attributes in RFC 14. Add a note that
these attributes may have no effect unless an accounting package is
installed and configured.