rfc42: add credit based flow control for channels
Problem: rfc42 does not specify anything about channel flow control,
so it is difficult to design clients and servers that operate without
dropping data.

Add credit based flow control to the protocol.

The client receives credit for writes to the subprocess via the
exec add-credit response.

The server receives credit for output to the client via the add-credit RPC.